+++ eqcorrscan/core/match_filter/party.py
@@ -27,7 +27,10 @@ import logging
 from os.path import join
 
 import numpy as np
-from obspy import Catalog, Stream, read_events
+from obspy import Catalog, Stream, read_events, UTCDateTime
+from obspy.core.event import (
+    StationMagnitude, Magnitude, ResourceIdentifier, WaveformStreamID,
+    CreationInfo, StationMagnitudeContribution)
 
 from eqcorrscan.core.match_filter.family import _write_family, _read_family
 from eqcorrscan.core.match_filter.matched_filter import (
@@ -41,6 +44,7 @@ from eqcorrscan.core.lag_calc import lag_calc
 from eqcorrscan.utils.catalog_utils import _get_origin
 from eqcorrscan.utils.findpeaks import decluster
 from eqcorrscan.utils.plotting import cumulative_detections
+from eqcorrscan.utils.mag_calc import relative_magnitude
 
 Logger = logging.getLogger(__name__)
 
@@ -742,7 +746,8 @@ class Party(object):
     def lag_calc(self, stream, pre_processed, shift_len=0.2, min_cc=0.4,
                  horizontal_chans=['E', 'N', '1', '2'], vertical_chans=['Z'],
                  cores=1, interpolate=False, plot=False, parallel=True,
-                 overlap='calculate', process_cores=None):
+                 overlap='calculate', process_cores=None,
+                 relative_magnitudes=False, **kwargs):
         """
         Compute picks based on cross-correlation alignment.
 
@@ -792,6 +797,14 @@ class Party(object):
         :param process_cores:
             Number of processes to use for pre-processing (if different to
             `cores`).
+        :type relative_magnitudes: bool
+        :param relative_magnitudes:
+            Whether to calculate relative magnitudes or not. See
+            :func:`eqcorrscan.utils.mag_calc.relative_magnitude` for more
+            information. Keyword arguments `noise_window`, `signal_window` and
+            `min_snr` can be passed as additional keyword arguments to pass
+            through to `eqcorrscan.utils.mag_calc.relative_magnitude`.
+
 
         :returns:
             Catalog of events with picks.  No origin information is included.
@@ -895,6 +908,56 @@ class Party(object):
                 for pick in event.picks:
                     pick.time += pre_pick
             catalog += temp_cat
+            if relative_magnitudes:
+                for event in temp_cat:
+                    det = [d for d in det_group
+                        if str(d.id) == str(event.resource_id)][0]
+                    corr_dict = {
+                        p.waveform_id.get_seed_string():
+                            float(p.comments[0].text.split("=")[-1])
+                        for p in event.picks}
+                    template = self.select(det.template_name).template
+                    try:
+                        t_mag = (
+                            template.event.preferred_magnitude() or
+                            template.event.megnitudes[0])
+                    except IndexError:
+                        Logger.info(
+                            "No template magnitude, relative magnitudes cannot"
+                            " be computed for {0}".format(event.resource_id))
+                        continue
+                    delta_mag = relative_magnitude(
+                        st1=template.st, st2=processed_stream,
+                        event1=template.event, event2=event,
+                        correlations=corr_dict, **kwargs)
+                    # Add station magnitudes
+                    sta_contrib = []
+                    av_mag = 0.0
+                    for seed_id, _delta_mag in delta_mag.items():
+                        sta_mag = StationMagnitude(
+                            mag=t_mag.mag + _delta_mag,
+                            magnitude_type=t_mag.magnitude_type,
+                            method_id=ResourceIdentifier("relative"),
+                            waveform_id=WaveformStreamID(seed_string=seed_id),
+                            creation_info=CreationInfo(
+                                author="EQcorrscan",
+                                creation_time=UTCDateTime()))
+                        event.station_magnitudes.append(sta_mag)
+                        sta_contrib.append(StationMagnitudeContribution(
+                            station_magnitude_id=sta_mag.resource_id,
+                            weight=1.))
+                        av_mag += sta_mag.mag
+                    av_mag /= len(delta_mag)
+                    # Compute average magnitude
+                    event.magnitudes.append(Magnitude(
+                        mag=av_mag, magnitude_type=t_mag.magnitude_type,
+                        method_id=ResourceIdentifier("relative"),
+                        station_count=len(delta_mag),
+                        evaluation_mode="manual",
+                        station_magnitude_contributions=sta_contrib,
+                        creation_info=CreationInfo(
+                                author="EQcorrscan",
+                                creation_time=UTCDateTime())))
         return catalog
 
     def get_catalog(self):